refactor(node): Extract BuyerRequestHandler for buyer/seller symmetry
Extract buyer-side request execution from node.ts into
buyer-request-handler.ts (310 lines):
- Streaming response assembly with timeout/abort/duration guards
- 402 payment negotiation retry loop
- Per-request SpendingAuth + cost tracking delegation to negotiator
- stripStreamingHeader + concatChunks utilities

node.ts sendRequest/sendRequestStream now delegate to the handler.
node.ts drops from 1375 → 1112 lines.

Buyer and seller are now fully separated:
- buyer-request-handler.ts — outbound request execution
- buyer-payment-negotiator.ts — payment state machine
- seller-request-handler.ts — inbound request routing
- seller-session-tracker.ts — session lifecycle + metering

Updated stream security and mux wiring tests to test the handler
directly instead of reaching into node internals.
