+++ packages/node/src/node.ts
@@ -7,12 +7,9 @@ import { loadOrCreateIdentity } from "./p2p/identity.js";
 import type { PeerId } from "./types/peer.js";
 import type { PeerInfo, TokenPricingUsdPerMillion } from "./types/peer.js";
 import { peerIdToAddress } from "./types/peer.js";
-import {
-  ANTSEED_STREAMING_RESPONSE_HEADER,
-  type SerializedHttpRequest,
-  type SerializedHttpResponse,
-  type SerializedHttpResponseChunk,
-  ANTSEED_SPENDING_AUTH_HEADER,
+import type {
+  SerializedHttpRequest,
+  SerializedHttpResponse,
 } from "./types/http.js";
 import type { ConnectionConfig } from "./types/connection.js";
 import { MeteringStorage } from "./metering/storage.js";
@@ -71,11 +68,18 @@ import { BuyerPaymentNegotiator } from "./payments/buyer-payment-negotiator.js";
 import { SellerPaymentManager, type SellerPaymentConfig } from "./payments/seller-payment-manager.js";
 import { IdentityClient } from "./payments/evm/identity-client.js";
 import { SellerRequestHandler } from "./seller-request-handler.js";
+import {
+  BuyerRequestHandler,
+  type RequestStreamCallbacks,
+  type RequestStreamResponseMetadata,
+  type RequestExecutionOptions,
+} from "./buyer-request-handler.js";
 
 export type { Provider, ProviderStreamCallbacks };
 export type { Router };
 export type { BuyerPaymentConfig };
 export type { SellerSessionSnapshot };
+export type { RequestStreamCallbacks, RequestStreamResponseMetadata, RequestExecutionOptions };
 
 export interface NodePaymentsConfig {
   /** Enable seller-side payment channels and automatic settlement. */
@@ -152,22 +156,6 @@ export interface NodeConfig {
   requireManualApproval?: boolean;
 }
 
-export interface RequestStreamResponseMetadata {
-  streaming: boolean;
-}
-
-export interface RequestStreamCallbacks {
-  onResponseStart?: (
-    response: SerializedHttpResponse,
-    metadata: RequestStreamResponseMetadata,
-  ) => void;
-  onResponseChunk?: (chunk: SerializedHttpResponseChunk) => void;
-}
-
-export interface RequestExecutionOptions {
-  signal?: AbortSignal;
-}
-
 export class AntseedNode extends EventEmitter {
   private _config: NodeConfig;
   private _identity: Identity | null = null;
@@ -196,6 +184,8 @@ export class AntseedNode extends EventEmitter {
   private _buyerPaymentManager: BuyerPaymentManager | null = null;
   /** Buyer-side payment negotiation (402 handling, SpendingAuth, cost tracking). */
   private _buyerNegotiator: BuyerPaymentNegotiator | null = null;
+  /** Buyer-side request execution (streaming, timeouts, 402 retry). */
+  private _buyerHandler: BuyerRequestHandler | null = null;
   /** Seller-side payment manager (initialized when seller has payment config). */
   private _sellerPaymentManager: SellerPaymentManager | null = null;
   /** Shared channel store for payment persistence. */
@@ -400,6 +390,7 @@ export class AntseedNode extends EventEmitter {
     this._identityClient = null;
     this._buyerPaymentManager = null;
     this._buyerNegotiator = null;
+    this._buyerHandler = null;
     this._sellerPaymentManager = null;
     this._sessionTracker = null;
     this._started = false;
@@ -464,8 +455,9 @@ export class AntseedNode extends EventEmitter {
   async connectToPeer(peer: PeerInfo): Promise<void> {
     const conn = await this._getOrCreateConnection(peer);
     this._getOrCreateMux(peer.peerId, conn);
-    if (this._buyerNegotiator) {
-      this._paymentMuxes.set(peer.peerId, this._buyerNegotiator.getOrCreatePaymentMux(peer.peerId, conn));
+    const negotiator = this._buyerNegotiator;
+    if (negotiator) {
+      this._paymentMuxes.set(peer.peerId, negotiator.getOrCreatePaymentMux(peer.peerId, conn));
     }
   }
 
@@ -474,7 +466,8 @@ export class AntseedNode extends EventEmitter {
     req: SerializedHttpRequest,
     options?: RequestExecutionOptions,
   ): Promise<SerializedHttpResponse> {
-    return this._sendRequestInternal(peer, req, undefined, options);
+    if (!this._buyerHandler) throw new Error("Node not started or not in buyer mode");
+    return this._buyerHandler.sendRequest(peer, req, undefined, options);
   }
 
   async sendRequestStream(
@@ -483,253 +476,8 @@ export class AntseedNode extends EventEmitter {
     callbacks: RequestStreamCallbacks,
     options?: RequestExecutionOptions,
   ): Promise<SerializedHttpResponse> {
-    return this._sendRequestInternal(peer, req, callbacks, options);
-  }
-
-  private async _sendRequestInternal(
-    peer: PeerInfo,
-    req: SerializedHttpRequest,
-    callbacks?: RequestStreamCallbacks,
-    options?: RequestExecutionOptions,
-  ): Promise<SerializedHttpResponse> {
-    if (!req.requestId || typeof req.requestId !== "string") {
-      throw new Error("requestId must be a non-empty string");
-    }
-    if (!this._connectionManager || !this._identity) {
-      throw new Error("Node not started");
-    }
-
-    const opName = callbacks ? "sendRequestStream" : "sendRequest";
-    debugLog(`[Node] ${opName} ${req.method} ${req.path} → peer ${peer.peerId.slice(0, 12)}... (reqId=${req.requestId.slice(0, 8)})`);
-
-    const conn = await this._getOrCreateConnection(peer);
-    debugLog(`[Node] Connection to ${peer.peerId.slice(0, 12)}... state=${conn.state}`);
-    const mux = this._getOrCreateMux(peer.peerId, conn);
-    const negotiator = this._buyerNegotiator;
-    if (negotiator) {
-      this._paymentMuxes.set(peer.peerId, negotiator.getOrCreatePaymentMux(peer.peerId, conn));
-    }
-
-    // Extract and strip x-antseed-spending-auth header if present (manual approval flow)
-    const externalSpendingAuth = req.headers[ANTSEED_SPENDING_AUTH_HEADER] ?? null;
-    if (externalSpendingAuth) {
-      const { [ANTSEED_SPENDING_AUTH_HEADER]: _, ...cleanHeaders } = req.headers;
-      req = { ...req, headers: cleanHeaders };
-    }
-
-    // If an external spending auth was provided, apply it before sending the request.
-    if (externalSpendingAuth && negotiator) {
-      debugLog(`[Node] Applying external spending auth for ${peer.peerId.slice(0, 12)}...`);
-      await negotiator.applyExternalSpendingAuth(peer, conn, externalSpendingAuth);
-    }
-
-    if (negotiator && !externalSpendingAuth) {
-      await negotiator.preparePreRequestAuth(peer, conn);
-    }
-
-    let startTime = Date.now();
-
-    const executeRequest = (): Promise<SerializedHttpResponse> => new Promise<SerializedHttpResponse>((resolve, reject) => {
-      const timeoutMs = this._config.requestTimeoutMs ?? 30_000;
-      const maxStreamBufferBytes = Math.max(1, this._config.maxStreamBufferBytes ?? 16 * 1024 * 1024);
-      const maxStreamDurationMs = Math.max(1, this._config.maxStreamDurationMs ?? 5 * 60_000);
-      const streamInitialResponseTimeoutMs = callbacks ? Math.max(timeoutMs, 90_000) : timeoutMs;
-      // Idle timeout for streaming: resets on each chunk so long-running
-      // streams (thinking models, large outputs) stay alive as long as
-      // data keeps flowing.
-      const streamIdleTimeoutMs = Math.max(timeoutMs, 60_000);
-      let settled = false;
-      let streamStarted = false;
-      let streamStartedAtMs = 0;
-      let streamBufferedBytes = 0;
-      let streamStartResponse: SerializedHttpResponse | null = null;
-      const streamChunks: Uint8Array[] = [];
-      let activeTimeout: ReturnType<typeof setTimeout> | null = null;
-      let activeTimeoutMs = streamInitialResponseTimeoutMs;
-      const abortSignal = options?.signal;
-      let abortListenerAttached = false;
-      let connectionStateListenerAttached = false;
-      const hasConnectionStateEvents =
-        typeof (conn as { on?: unknown }).on === "function"
-        && typeof (conn as { off?: unknown }).off === "function";
-
-      const cleanupAbortListener = (): void => {
-        if (abortSignal && abortListenerAttached) {
-          abortSignal.removeEventListener("abort", onAbort);
-          abortListenerAttached = false;
-        }
-      };
-      const cleanupConnectionListener = (): void => {
-        if (!connectionStateListenerAttached) return;
-        conn.off("stateChange", onConnectionStateChange);
-        connectionStateListenerAttached = false;
-      };
-      const onConnectionStateChange = (state: ConnectionState): void => {
-        if (settled) return;
-        if (state !== ConnectionState.Closed && state !== ConnectionState.Failed) {
-          return;
-        }
-        settled = true;
-        if (activeTimeout) clearTimeout(activeTimeout);
-        cleanupAbortListener();
-        cleanupConnectionListener();
-        mux.cancelProxyRequest(req.requestId);
-        reject(new Error(`Connection to ${peer.peerId} ${state.toLowerCase()} during request ${req.requestId}`));
-      };
-
-      const onAbort = (): void => {
-        if (settled) return;
-        settled = true;
-        if (activeTimeout) clearTimeout(activeTimeout);
-        cleanupAbortListener();
-        cleanupConnectionListener();
-        debugWarn(`[Node] Request ${req.requestId.slice(0, 8)} aborted by caller`);
-        mux.cancelProxyRequest(req.requestId);
-        reject(new Error(`Request ${req.requestId} aborted`));
-      };
-
-      if (abortSignal) {
-        if (abortSignal.aborted) {
-          onAbort();
-          return;
-        }
-        abortSignal.addEventListener("abort", onAbort, { once: true });
-        abortListenerAttached = true;
-      }
-      if (hasConnectionStateEvents) {
-        conn.on("stateChange", onConnectionStateChange);
-        connectionStateListenerAttached = true;
-      }
-
-      const resetTimeout = (ms: number): void => {
-        if (activeTimeout) clearTimeout(activeTimeout);
-        activeTimeoutMs = ms;
-        activeTimeout = setTimeout(() => {
-          if (settled) return;
-          settled = true;
-          cleanupAbortListener();
-          cleanupConnectionListener();
-          debugWarn(
-            `[Node] Request ${req.requestId.slice(0, 8)} timed out after ${Date.now() - startTime}ms `
-            + `(timeout=${activeTimeoutMs}ms, stream=${callbacks ? "true" : "false"}, streamStarted=${streamStarted ? "true" : "false"}, buffered=${streamBufferedBytes}b)`,
-          );
-          mux.cancelProxyRequest(req.requestId);
-          reject(new Error(`Request ${req.requestId} timed out`));
-        }, ms);
-      };
-
-      // Initial timeout: wait for the first response frame.
-      resetTimeout(streamInitialResponseTimeoutMs);
-
-      const finish = (response: SerializedHttpResponse): void => {
-        if (settled) return;
-        settled = true;
-        if (activeTimeout) clearTimeout(activeTimeout);
-        cleanupAbortListener();
-        cleanupConnectionListener();
-        const cleaned = this._stripStreamingHeader(response);
-        debugLog(`[Node] Response for ${req.requestId.slice(0, 8)}: status=${cleaned.statusCode} (${Date.now() - startTime}ms, ${cleaned.body.length}b)`);
-        resolve(cleaned);
-      };
-
-      const fail = (error: Error): void => {
-        if (settled) return;
-        settled = true;
-        if (activeTimeout) clearTimeout(activeTimeout);
-        cleanupAbortListener();
-        cleanupConnectionListener();
-        reject(error);
-      };
-
-      mux.sendProxyRequest(
-        req,
-        (response: SerializedHttpResponse, metadata) => {
-          if (settled) return;
-          if (metadata.streamingStart) {
-            streamStarted = true;
-            streamStartedAtMs = Date.now();
-            streamBufferedBytes = 0;
-            streamStartResponse = this._stripStreamingHeader(response);
-            debugLog(`[Node] Stream started for ${req.requestId.slice(0, 8)}; idle-timeout=${streamIdleTimeoutMs}ms`);
-            // Switch to streaming idle timeout: resets on each chunk.
-            resetTimeout(streamIdleTimeoutMs);
-            callbacks?.onResponseStart?.(streamStartResponse, { streaming: true });
-            return;
-          }
-
-          callbacks?.onResponseStart?.(this._stripStreamingHeader(response), { streaming: false });
-          finish(response);
-        },
-        (chunk) => {
-          if (settled) return;
-          if (!streamStarted) return;
-
-          // Reset idle timeout on each chunk so streaming stays alive.
-          resetTimeout(streamIdleTimeoutMs);
-
-          if (Date.now() - streamStartedAtMs > maxStreamDurationMs) {
-            mux.cancelProxyRequest(req.requestId);
-            fail(new Error(`Stream ${req.requestId} exceeded max duration (${maxStreamDurationMs}ms)`));
-            return;
-          }
-
-          callbacks?.onResponseChunk?.(chunk);
-
-          if (chunk.data.length > 0) {
-            if (callbacks?.onResponseChunk) {
-              // Streaming mode: chunks already delivered to caller via callback.
-              // Track byte count for the debug timeout log only — do not
-              // enforce maxStreamBufferBytes so large streams aren't rejected.
-              streamBufferedBytes += chunk.data.length;
-              streamChunks.push(chunk.data);
-            } else {
-              // Non-streaming: accumulate chunks for the final response body.
-              const nextBufferedBytes = streamBufferedBytes + chunk.data.length;
-              if (nextBufferedBytes > maxStreamBufferBytes) {
-                mux.cancelProxyRequest(req.requestId);
-                fail(new Error(`Stream ${req.requestId} exceeded max buffered size (${maxStreamBufferBytes} bytes)`));
-                return;
-              }
-              streamBufferedBytes = nextBufferedBytes;
-              streamChunks.push(chunk.data);
-            }
-          }
-
-          if (!chunk.done) return;
-
-          if (!streamStartResponse) {
-            fail(new Error(`Stream ${req.requestId} ended before response start`));
-            return;
-          }
-
-          finish({
-            ...streamStartResponse,
-            body: concatChunks(streamChunks),
-          });
-        },
-      );
-    });
-
-    // Execute the request. If we get a 402 and payment negotiation is needed,
-    // wait for the seller's PaymentRequired message, negotiate, and retry.
-    const response = await executeRequest();
-
-    if (response.statusCode === 402 && negotiator && !externalSpendingAuth) {
-      const result = await negotiator.handle402(response, peer, conn, req);
-      if (result.action === 'return') return result.response;
-      // Retry after successful negotiation
-      startTime = Date.now();
-      return executeRequest();
-    }
-
-    // Track response cost for buyer-side per-request auth verification
-    if (negotiator) {
-      negotiator.estimateCostFromResponse(peer, response);
-      negotiator.parseCostHeaders(peer.peerId, response);
-      negotiator.recordResponseContent(peer.peerId, req.body, response.body, Date.now() - startTime);
-    }
-
-    return response;
+    if (!this._buyerHandler) throw new Error("Node not started or not in buyer mode");
+    return this._buyerHandler.sendRequest(peer, req, callbacks, options);
   }
 
   private _createDHTConfig(port: number, bootstrapNodes: Array<{ host: string; port: number }>): DHTNodeConfig {
@@ -1048,6 +796,21 @@ export class AntseedNode extends EventEmitter {
       }
     }
 
+    // Create buyer request handler
+    this._buyerHandler = new BuyerRequestHandler(
+      {
+        requestTimeoutMs: this._config.requestTimeoutMs,
+        maxStreamBufferBytes: this._config.maxStreamBufferBytes,
+        maxStreamDurationMs: this._config.maxStreamDurationMs,
+      },
+      {
+        negotiator: this._buyerNegotiator,
+        getConnection: (peer) => this._getOrCreateConnection(peer),
+        getMux: (peerId, conn) => this._getOrCreateMux(peerId, conn),
+        registerPaymentMux: (peerId, pmux) => this._paymentMuxes.set(peerId, pmux),
+      },
+    );
+
     debugLog(`[Node] Buyer ready — DHT running on port ${this._dht!.getPort()}`);
   }
 
@@ -1083,19 +846,6 @@ export class AntseedNode extends EventEmitter {
     this._muxes.set(buyerPeerId, mux);
     this._wireConnection(conn, buyerPeerId);
     this.emit("connection", conn);
-  }
-
-  private _stripStreamingHeader(response: SerializedHttpResponse): SerializedHttpResponse {
-    if (response.headers[ANTSEED_STREAMING_RESPONSE_HEADER] !== "1") {
-      return response;
-    }
-
-    const headers = { ...response.headers };
-    delete headers[ANTSEED_STREAMING_RESPONSE_HEADER];
-    return {
-      ...response,
-      headers,
-    };
   }
 
   private async _initializePayments(dataDir: string): Promise<void> {
@@ -1360,16 +1110,3 @@ function parsePeerAddress(address: string): { host: string; port: number } {
   return { host: parts[0]!, port: parseInt(parts[1] ?? "6882", 10) };
 }
 
-function concatChunks(chunks: Uint8Array[]): Uint8Array {
-  if (chunks.length === 0) return new Uint8Array(0);
-  if (chunks.length === 1) return chunks[0]!;
-
-  const totalLength = chunks.reduce((sum, chunk) => sum + chunk.length, 0);
-  const output = new Uint8Array(totalLength);
-  let offset = 0;
-  for (const chunk of chunks) {
-    output.set(chunk, offset);
-    offset += chunk.length;
-  }
-  return output;
-}
