+++ packages/node/tests/node-payment-mux-wiring.test.ts
@@ -1,31 +1,15 @@
 import { afterEach, describe, expect, it, vi } from 'vitest';
-import { AntseedNode, type NodeConfig } from '../src/node.js';
+import { BuyerRequestHandler } from '../src/buyer-request-handler.js';
 import type { SerializedHttpRequest } from '../src/types/http.js';
 import type { SerializedHttpResponse } from '../src/types/http.js';
-import type { PeerInfo } from '../src/types/peer.js';
+import type { PeerInfo, PeerId } from '../src/types/peer.js';
 
-function createNode(config: Partial<NodeConfig> = {}): AntseedNode {
-  const node = new AntseedNode({
-    role: 'buyer',
-    ...config,
-  });
-
-  (node as any)._identity = {
-    peerId: 'a'.repeat(40),
-    privateKey: new Uint8Array(32),
-    publicKey: new Uint8Array(32),
-  };
-  (node as any)._connectionManager = {};
-  return node;
-}
-
-describe('AntseedNode buyer payment mux wiring', () => {
+describe('BuyerRequestHandler payment mux wiring', () => {
   afterEach(() => {
     vi.restoreAllMocks();
   });
 
   it('creates buyer payment mux before sending outbound requests when payments are enabled', async () => {
-    const node = createNode();
     const peer = { peerId: 'b'.repeat(40) } as PeerInfo;
     const request: SerializedHttpRequest = {
       requestId: 'req-payment-mux',
@@ -49,20 +33,27 @@ describe('AntseedNode buyer payment mux wiring', () => {
     });
 
     const getOrCreatePaymentMux = vi.fn().mockReturnValue({});
-    (node as any)._buyerNegotiator = {
-      getOrCreatePaymentMux: getOrCreatePaymentMux,
-      preparePreRequestAuth: vi.fn(),
-      estimateCostFromResponse: vi.fn(),
-      parseCostHeaders: vi.fn(),
-      recordResponseContent: vi.fn(),
-    };
-    (node as any)._getOrCreateConnection = vi.fn(async () => conn);
-    (node as any)._getOrCreateMux = vi.fn(() => ({
-      sendProxyRequest,
-      cancelProxyRequest: vi.fn(),
-    }));
+    const registerPaymentMux = vi.fn();
+    const handler = new BuyerRequestHandler(
+      {},
+      {
+        negotiator: {
+          getOrCreatePaymentMux,
+          preparePreRequestAuth: vi.fn(),
+          estimateCostFromResponse: vi.fn(),
+          parseCostHeaders: vi.fn(),
+          recordResponseContent: vi.fn(),
+        } as any,
+        getConnection: vi.fn(async () => conn) as any,
+        getMux: vi.fn(() => ({
+          sendProxyRequest,
+          cancelProxyRequest: vi.fn(),
+        })) as any,
+        registerPaymentMux,
+      },
+    );
 
-    await (node as any)._sendRequestInternal(peer, request, undefined);
+    await handler.sendRequest(peer, request);
 
     expect(getOrCreatePaymentMux).toHaveBeenCalledWith(peer.peerId, conn);
     expect(sendProxyRequest).toHaveBeenCalledOnce();
@@ -72,7 +63,6 @@ describe('AntseedNode buyer payment mux wiring', () => {
   });
 
   it('re-negotiates on 402 even when the peer was already locked', async () => {
-    const node = createNode();
     const peer = { peerId: 'b'.repeat(40) } as PeerInfo;
     const request: SerializedHttpRequest = {
       requestId: 'req-relock',
@@ -111,21 +101,27 @@ describe('AntseedNode buyer payment mux wiring', () => {
     });
 
     const handle402 = vi.fn(async () => ({ action: 'retry' as const }));
-    (node as any)._buyerNegotiator = {
-      getOrCreatePaymentMux: vi.fn().mockReturnValue({}),
-      preparePreRequestAuth: vi.fn(),
-      handle402,
-      estimateCostFromResponse: vi.fn(),
-      parseCostHeaders: vi.fn(),
-      recordResponseContent: vi.fn(),
-    };
-    (node as any)._getOrCreateConnection = vi.fn(async () => conn);
-    (node as any)._getOrCreateMux = vi.fn(() => ({
-      sendProxyRequest,
-      cancelProxyRequest: vi.fn(),
-    }));
+    const handler = new BuyerRequestHandler(
+      {},
+      {
+        negotiator: {
+          getOrCreatePaymentMux: vi.fn().mockReturnValue({}),
+          preparePreRequestAuth: vi.fn(),
+          handle402,
+          estimateCostFromResponse: vi.fn(),
+          parseCostHeaders: vi.fn(),
+          recordResponseContent: vi.fn(),
+        } as any,
+        getConnection: vi.fn(async () => conn) as any,
+        getMux: vi.fn(() => ({
+          sendProxyRequest,
+          cancelProxyRequest: vi.fn(),
+        })) as any,
+        registerPaymentMux: vi.fn(),
+      },
+    );
 
-    const response = await (node as any)._sendRequestInternal(peer, request, undefined);
+    const response = await handler.sendRequest(peer, request);
 
     expect(response.statusCode).toBe(200);
     expect(handle402).toHaveBeenCalledOnce();
